+++ src/main/java/dev/shadowsoffire/placebo/json/JsonUtil.java
@@ -69,6 +69,7 @@ public class JsonUtil {
         return regObj;
     }
 
+    @Deprecated
     public static <T> Object makeSerializer(Registry<T> reg) {
         return new SDS<>(reg);
     }
@@ -76,6 +77,7 @@ public class JsonUtil {
     /**
      * Short for Serializer/Deserializer
      */
+    @Deprecated
     private static class SDS<T> implements com.google.gson.JsonDeserializer<T>, com.google.gson.JsonSerializer<T> {
 
         private final Registry<T> reg;
@@ -98,22 +100,27 @@ public class JsonUtil {
 
     }
 
+    @Deprecated
     public static interface JsonSerializer<V> {
         public JsonObject write(V src);
     }
 
+    @Deprecated
     public static interface JsonDeserializer<V> {
         public V read(JsonObject json);
     }
 
+    @Deprecated
     public static interface NetSerializer<V> {
         public void write(V src, FriendlyByteBuf buf);
     }
 
+    @Deprecated
     public static interface NetDeserializer<V> {
         public V read(FriendlyByteBuf buf);
     }
 
+    @Deprecated
     private static record SDS2<T>(com.google.gson.JsonDeserializer<T> jds, com.google.gson.JsonSerializer<T> js) implements com.google.gson.JsonDeserializer<T>, com.google.gson.JsonSerializer<T> {
 
         @Override
@@ -128,6 +135,7 @@ public class JsonUtil {
 
     }
 
+    @Deprecated
     public static <T> Object makeSerializer(com.google.gson.JsonDeserializer<T> jds, com.google.gson.JsonSerializer<T> js) {
         return new SDS2<>(jds, js);
     }