+++ src/main/java/dev/shadowsoffire/placebo/json/RandomAttributeModifier.java
@@ -22,54 +22,59 @@ import java.util.Objects;
 import java.util.Random;
 import java.util.UUID;
 
-public class RandomAttributeModifier {
+public record RandomAttributeModifier(Attribute attribute, Operation op, StepFunction value, UUID id) {
 
     public static Codec<RandomAttributeModifier> CODEC = RecordCodecBuilder.create(inst -> inst
-        .group(
-            BuiltInRegistries.ATTRIBUTE.byNameCodec().fieldOf("attribute").forGetter(a -> a.attribute),
-            PlaceboCodecs.enumCodec(Operation.class).fieldOf("operation").forGetter(a -> a.op),
-            StepFunction.CODEC.fieldOf("value").forGetter(a -> a.value))
-        .apply(inst, RandomAttributeModifier::new));
+            .group(
+                    BuiltInRegistries.ATTRIBUTE.byNameCodec().fieldOf("attribute").forGetter(a -> a.attribute),
+                    PlaceboCodecs.enumCodec(Operation.class).fieldOf("operation").forGetter(a -> a.op),
+                    StepFunction.CODEC.fieldOf("value").forGetter(a -> a.value))
+            .apply(inst, RandomAttributeModifier::new));
 
-    protected final Attribute attribute;
-    protected final Operation op;
-    protected final StepFunction value;
-    protected final UUID id;
+    public static Codec<RandomAttributeModifier> CONSTANT_CODEC = RecordCodecBuilder.create(inst -> inst
+            .group(
+                    BuiltInRegistries.ATTRIBUTE.byNameCodec().fieldOf("attribute").forGetter(a -> a.attribute),
+                    PlaceboCodecs.enumCodec(Operation.class).fieldOf("operation").forGetter(a -> a.op),
+                    StepFunction.CONSTANT_CODEC.fieldOf("value").forGetter(a -> a.value))
+            .apply(inst, RandomAttributeModifier::new));
 
     /**
-     * Creates a Chanced Effect Instance.
+     * Creates a Random Attribute Modifier. A UUID will be randomly generated.
      *
-     * param chance The chance this potion is received.
-     * param effect The effect.
-     * param amp    A random range of possible amplifiers.
+     * @param attribute The attribute the generated modifier will be applicable to.
+     * @param op        The operation of the generated modifier.
+     * @param value     The value range for the generated modifier.
      */
     public RandomAttributeModifier(Attribute attribute, Operation op, StepFunction value) {
-        this.attribute = attribute;
-        this.op = op;
-        this.value = value;
-        Random rand = new Random();
-        rand.setSeed(Objects.hash(attribute, op, value));
-        this.id = new UUID(rand.nextLong(), rand.nextLong());
+        this(attribute, op, value, UUID.randomUUID());
     }
 
     public void apply(RandomSource rand, LivingEntity entity) {
         if (entity == null) throw new RuntimeException("Attempted to apply a random attribute modifier to a null entity!");
-        AttributeModifier modif = this.genModifier(rand);
+        AttributeModifier modif = this.create(rand);
         AttributeInstance inst = entity.getAttribute(this.attribute);
         if (inst == null) {
             Placebo.LOGGER
-                .trace(String.format("Attempted to apply a random attribute modifier to an entity (%s) that does not have that attribute (%s)!", EntityType.getKey(entity.getType()), BuiltInRegistries.ATTRIBUTE.getKey(this.attribute)));
+                    .trace(String.format("Attempted to apply a random attribute modifier to an entity (%s) that does not have that attribute (%s)!", EntityType.getKey(entity.getType()), BuiltInRegistries.ATTRIBUTE.getKey(this.attribute)));
             return;
         }
         inst.addPermanentModifier(modif);
     }
 
-    public AttributeModifier genModifier(RandomSource rand) {
+    public AttributeModifier create(RandomSource rand) {
         return new AttributeModifier(this.id, "placebo_random_modifier_" + this.attribute.getDescriptionId(), this.value.get(rand.nextFloat()), this.op);
     }
 
-    public AttributeModifier genModifier(String name, RandomSource rand) {
+    public AttributeModifier create(String name, RandomSource rand) {
         return new AttributeModifier(name, this.value.get(rand.nextFloat()), this.op);
+    }
+
+    public AttributeModifier createDeterministic() {
+        return new AttributeModifier(this.id, "placebo_random_modifier_" + this.attribute.getDescriptionId(), this.value.min(), this.op);
+    }
+
+    public AttributeModifier createDeterministic(String name) {
+        return new AttributeModifier(this.id, name, this.value.min(), this.op);
     }
 
     public Attribute getAttribute() {